requirements.txt
requests
beautifulsoup4
telebot
telegram_text
python-decouple